Committing the program.
As it currently stands, this program is not good. It was written at the tail end of 2016, and I've improved my ability to code - particularly in C++ - by a lot since then. I'm committing this so that it's always available, and that others can fork it if there's actually a reason to.

P.S. This project file is configured to use Visual Studio 2017, if you have an older version, this should still work, but you'll need to change the toolset in project properties.

Signed-off-by: Drew Ritter <dritter@razelon.com>